refactor(phase-4.2c): Finalize util.py refactoring - 86% reduction
COMPLETED: util.py module refactoring

Changes:
- Replaced util.py (1,042 lines) with re-export module (143 lines)
- 86% line reduction: 1,042  143 lines
- Backed up original as util_old.py for reference

New util.py structure:
- Imports all 45 functions from util_modules subpackages
- Provides backward compatibility via re-exports
- Clean __all__ export list for API clarity
- Comprehensive docstring explaining new organization

Re-exported functions (45 total):
 xml_helpers (20): read_bool, parse_bool, safe_read, read_property, create_property_node, etc.
 calibration (3): create_calibration_function, with_center_calibration, no_center_calibration
 path_utils (2): resource_path, userprofile_path
 file_operations (1): FileWatcher
 misc (17): clamp, rad2deg, deg2rad, hat_*_direction, format_name, etc.

Benefits achieved:
 Single Responsibility Principle (SRP) - each module has one focus
 Easier testing - functions grouped by concern
 Better discoverability - clear module names
 Reduced coupling - cleaner dependencies
 Backward compatibility - existing imports still work
 Maintainability - 5 focused files instead of 1 monolith

Module breakdown:
- gremlin/util_modules/xml_helpers.py: 700 lines (XML/property handling)
- gremlin/util_modules/calibration.py: 106 lines (axis calibration)
- gremlin/util_modules/path_utils.py: 52 lines (path resolution)
- gremlin/util_modules/file_operations.py: 63 lines (file monitoring)
- gremlin/util_modules/misc.py: 340 lines (miscellaneous utilities)
- gremlin/util.py: 143 lines (re-exports)

Total: ~1,404 lines (organized) vs 1,042 lines (monolithic)

Next: Split remaining large modules (profile.py, ui/profile.py)

Refs: #phase-4 #clean-code #solid #srp #refactoring-complete
